requirements.txt
git+https://github.com/oatsu-gh/nnsvs@kuresampler
git+https://github.com/oatsu-gh/PyUtauCli@kuresampler
git+https://github.com/oatsu-gh/PyRwu@kuresampler
git+https://github.com/oatsu-gh/PyWavTool@kuresampler
git+https://github.com/oatsu-gh/ParallelWaveGAN@kuresampler
git+https://github.com/oatsu-gh/HN-UnifiedSourceFilterGAN@kuresampler
numpy<2
omegaconf
light-the-torch
colored-traceback
colorlog
# torch should be installed manually according to the environment
